requirements.txt
flask
PySerial
